Moved util/ballot.js to functions, imports updated
